// src/types/enums.ts
export enum UserRole {
    GPO = "GPO",
    VENDOR = "VENDOR"
}

export enum ContractStatus {
    ACTIVE = "ACTIVE",
    COMPLETED = "COMPLETED",
    TERMINATED = "TERMINATED"
}

export enum MilestoneStatus {
    NOT_STARTED = "NOT_STARTED",
    IN_PROGRESS = "IN_PROGRESS",
    COMPLETED = "COMPLETED",
    DELAYED = "DELAYED"
} 